docling-rag-project/requirements.txt
docling>=1.0.0
cohere>=5.0.0
opensearch-py>=2.3.0
gremlinpython>=3.6.0
boto3>=1.28.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
